Remove forced mode and the creation of backups
Changes are always stored in disk. However, now the strategy is being
more resilient to resumes. Changes that have been propagated to DSpace
are considered correct and saved. If the program aborts, the saved state
should allow resuming on the point of failure.

Backups are not kept since they were creating a lot of trash files that
rarely were reused o checked.